use offerings in bootstrap data, update typescript use
