+++ tutor/scripts/test-server/backend/db.json
@@ -3,6 +3,7 @@
   "ui-settings": {
     "ui_settings": {}
   },
+  "suggest": {},
   "updates": {},
   "onboarding": {},
   "tours": {},
