+++ tutor/scripts/test-server/backend/suggest.js
@@ -1,0 +1,13 @@
+const Factory = require('object-factory-bot');
+
+module.exports = {
+
+  setRole(){ },
+
+  route(server) {
+    server.post('/api/user/suggest', (req, res) => {
+      return res.json({});
+    });
+  },
+
+};
