+++ tutor/specs/e2e/my-courses.e2e.ts
@@ -14,4 +14,22 @@ describe('My Courses', () => {
         await expect(page).toHaveSelector('testEl=new-teacher-screen')
         await expect(page).not.toHaveSelector('testEl=existing-teacher-screen', { timeout: 10 })
     })
+
+    it.only('allows a new teacher select and suggest subjects', async () => {
+        await visitPage(page, '/courses')
+        await expect(page).toHaveSelector('testEl=existing-teacher-screen')
+        await page.evaluate(() => {
+            window._MODELS.store.dispatch({ type: 'bootstrap', payload: { courses: [] } })
+        })
+        await page.type('testEl=input-suggested-subject', 'test')
+        await page.click('testEl=submit-suggested-subject')
+        await expect(page).not.toHaveSelector('testEl=existing-teacher-screen', { timeout: 100 })
+        expect(
+            await page.evaluate(() => document.location.search)
+        ).toContain('onboarding=2')
+        await page.click('testEl=back-to-select')
+        expect(
+            await page.evaluate(() => document.location.search)
+        ).toContain('onboarding=0')
+    })
 })
