+++ tutor/src/screens/my-courses/index.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux'
 import Dashboard from './dashboard'
 import { useHasAnyCourses } from '../../store/courses'
 import { getOfferings } from '../../store/api'
-import NewUser from './new-user'
+import NewTeacher from './new-teacher'
 
 const MyCourses: React.FC = (props) => {
   const d = useDispatch()
@@ -15,7 +15,7 @@ const MyCourses: React.FC = (props) => {
     return <Dashboard />
   }
   return (
-    <NewUser history={props.history} data-test-id="new-teacher-screen" />
+    <NewTeacher history={props.history} data-test-id="new-teacher-screen" />
   )
 }
